[util] Move face/font options into face/font-options.hh
